requirements.txt
# Core LLM & RAG Framework
langchain>=0.1.0
langchain-community>=0.1.0
langchain-openai>=0.1.0

# Vector Database
chromadb>=0.4.0

# API Framework
fastapi>=0.100.0
uvicorn>=0.25.0
pydantic>=2.0.0

# Document Processing
pypdf>=4.0.0
python-docx>=0.8.0
python-multipart>=0.0.6

# Embeddings & Models
openai>=1.0.0

# Evaluation (RAGAS)
ragas>=0.0.60

# Data & Utils
python-dotenv>=1.0.0
requests>=2.30.0
numpy>=1.24.0
pandas>=2.0.0

# Logging & Monitoring
python-json-logger>=2.0.0

# Development
pytest>=7.0.0
black>=24.0.0
flake8>=7.0.0
